requirements.txt
requests==2.31.0
numpy==1.26.4
pandas==2.2.2
psycopg2-binary==2.9.9
pygame
yfinance
matplotlib
scipy
ta
dotenv
